requirements.txt
azure-cognitiveservices-speech==1.16.0
firebase-admin==4.5.3
google-cloud-firestore==2.1.0
openai==0.27.0